refactor(gui): split LiveSessionTab into by-concern partials
The 2,649-line monolith becomes a 383-line core (shared state/colours +
Render orchestration) plus six partials, mirroring the ImageRendererBase
by-concern convention: .Input (mouse + keyboard), .Strips (top strip,
timeline, bottom strip, abort-confirm), .Panels (per-OTA panels, cooling
sparkline, exposure log), .Charts (guide graph, V-curve), .Preview
(mini-viewer toolbar, preview timeline, capture controls, mount section),
and .Polar (polar-alignment mode).

Pure verbatim method moves -- verified by a line-multiset diff against
HEAD (only the class declaration changed, to partial; every added line is
per-file usings/banner/brace boilerplate). 0-warning build;
PolarAlignmentSessionTests (the suite touching this class) green.
